Update omnidirectional-robot-car.ino
Fixed velocity and motor speed calculations by computing velocities relative to the max possible for each individual wheel as well as for all four wheels together. This is necessary because the range of values for the joystick X and Y position outline a square about the origin, but the calculations for translating the X and Y values of the joystick to four individual velocity values assume that the joystick X and Y positions outline a circle about the origin.